Module nom8::bytes

Parsers recognizing bytes streams

Modules

complete

Parsers recognizing bytes streams, complete input version

streaming

Parsers recognizing bytes streams, streaming version

Functions

any

Matches one token

escaped

Matches a byte string with escaped characters.

escaped_transform

Matches a byte string with escaped characters.

none_of

Returns a token that does not match the pattern

one_of

Returns a token that matches the pattern

tag

Recognizes a pattern

tag_no_case

Recognizes a case insensitive pattern.

take

Returns an input slice containing the first N input elements (Input[..N]).

take_till

Returns the longest input slice (if any) till a pattern is met.

take_till1

Returns the longest (at least 1) input slice till a pattern is met.

take_until

Returns the input slice up to the first occurrence of the pattern.

take_until1

Returns the non empty input slice up to the first occurrence of the pattern.

take_while

Returns the longest input slice (if any) that matches the pattern

take_while1

Returns the longest (at least 1) input slice that matches the pattern

take_while_m_n

Returns the longest (m <= len <= n) input slice that matches the pattern