Function nom8::bytes::complete::take_till

pub fn take_till<T, Input, Error: ParseError<Input>>(
    list: T
) -> impl Fn(Input) -> IResult<Input, <Input as IntoOutput>::Output, Error>where
    Input: InputTakeAtPosition,
    Input: IntoOutput,
    T: FindToken<<Input as InputTakeAtPosition>::Item>,

👎Deprecated since 8.0.0:
Replaced with nom8::bytes::take_till

Returns the longest input slice (if any) till a predicate is met.

The parser will return the longest slice till the given predicate (a function that takes the input and returns a bool).

Example

use nom8::bytes::complete::take_till;

fn till_colon(s: &str) -> IResult<&str, &str> {
  take_till(|c| c == ':')(s)
}

assert_eq!(till_colon("latin:123"), Ok((":123", "latin")));
assert_eq!(till_colon(":empty matched"), Ok((":empty matched", ""))); //allowed
assert_eq!(till_colon("12345"), Ok(("", "12345")));
assert_eq!(till_colon(""), Ok(("", "")));

WARNING: Deprecated, replaced with nom8::bytes::take_till