Function nom8::bytes::complete::take_while

pub fn take_while<T, Input, Error: ParseError<Input>>(
    list: T
) -> impl Fn(Input) -> IResult<Input, <Input as IntoOutput>::Output, Error>where
    Input: InputTakeAtPosition + IntoOutput,
    T: FindToken<<Input as InputTakeAtPosition>::Item>,

👎Deprecated since 8.0.0: Replaced with nom8::bytes::take_while

Returns the longest input slice (if any) that matches the predicate.

The parser will return the longest slice that matches the given predicate (a function that takes the input and returns a bool).

Example

use nom8::bytes::complete::take_while;
use nom8::input::AsChar;

fn alpha(s: &[u8]) -> IResult<&[u8], &[u8]> {
  take_while(AsChar::is_alpha)(s)
}

assert_eq!(alpha(b"latin123"), Ok((&b"123"[..], &b"latin"[..])));
assert_eq!(alpha(b"12345"), Ok((&b"12345"[..], &b""[..])));
assert_eq!(alpha(b"latin"), Ok((&b""[..], &b"latin"[..])));
assert_eq!(alpha(b""), Ok((&b""[..], &b""[..])));

WARNING: Deprecated, replaced with nom8::bytes::take_while