nom8::bytes::complete
Function take_while_m_n

pub fn take_while_m_n<T, Input, Error: ParseError<Input>>(
    m: usize,
    n: usize,
    list: T,
) -> impl Fn(Input) -> IResult<Input, <Input as IntoOutput>::Output, Error>
where
    Input: InputTake + InputIter + InputLength + Slice<RangeFrom<usize>> + IntoOutput,
    T: FindToken<<Input as InputIter>::Item>,

👎Deprecated since 8.0.0: Replaced with nom8::bytes::take_while_m_n

Returns the longest (m <= len <= n) input slice that matches the predicate.

The parser will return the longest slice that matches the given predicate (a function that takes the input and returns a bool).

It will return an Err::Error((_, ErrorKind::TakeWhileMN)) if the pattern wasn’t met or is out of range (m <= len <= n).

Example

use nom8::bytes::complete::take_while_m_n;
use nom8::input::AsChar;

fn short_alpha(s: &[u8]) -> IResult<&[u8], &[u8]> {
  take_while_m_n(3, 6, AsChar::is_alpha)(s)
}

assert_eq!(short_alpha(b"latin123"), Ok((&b"123"[..], &b"latin"[..])));
assert_eq!(short_alpha(b"lengthy"), Ok((&b"y"[..], &b"length"[..])));
assert_eq!(short_alpha(b"latin"), Ok((&b""[..], &b"latin"[..])));
assert_eq!(short_alpha(b"ed"), Err(Err::Error(Error::new(&b"ed"[..], ErrorKind::TakeWhileMN))));
assert_eq!(short_alpha(b"12345"), Err(Err::Error(Error::new(&b"12345"[..], ErrorKind::TakeWhileMN))));

WARNING: Deprecated, replaced with nom8::bytes::take_while_m_n