nom8::bytes
Function tag_no_case

pub fn tag_no_case<T, Input, Error: ParseError<Input>, const STREAMING: bool>(
    tag: T,
) -> impl Fn(Input) -> IResult<Input, <Input as IntoOutput>::Output, Error>
where
    Input: InputTake + InputLength + Compare<T> + InputIsStreaming<STREAMING> + IntoOutput,
    T: InputLength + Clone,

Recognizes a case insensitive pattern.

The input data will be compared to the tag combinator’s argument and will return the part of the input that matches the argument with no regard to case.

It will return Err(Err::Error((_, ErrorKind::Tag))) if the input doesn’t match the pattern.

Example

use nom8::bytes::tag_no_case;

fn parser(s: &str) -> IResult<&str, &str> {
  tag_no_case("hello")(s)
}

assert_eq!(parser("Hello, World!"), Ok((", World!", "Hello")));
assert_eq!(parser("hello, World!"), Ok((", World!", "hello")));
assert_eq!(parser("HeLlO, World!"), Ok((", World!", "HeLlO")));
assert_eq!(parser("Something"), Err(Err::Error(Error::new("Something", ErrorKind::Tag))));
assert_eq!(parser(""), Err(Err::Error(Error::new("", ErrorKind::Tag))));

use nom8::bytes::tag_no_case;

fn parser(s: Streaming<&str>) -> IResult<Streaming<&str>, &str> {
  tag_no_case("hello")(s)
}

assert_eq!(parser(Streaming("Hello, World!")), Ok((Streaming(", World!"), "Hello")));
assert_eq!(parser(Streaming("hello, World!")), Ok((Streaming(", World!"), "hello")));
assert_eq!(parser(Streaming("HeLlO, World!")), Ok((Streaming(", World!"), "HeLlO")));
assert_eq!(parser(Streaming("Something")), Err(Err::Error(Error::new(Streaming("Something"), ErrorKind::Tag))));
assert_eq!(parser(Streaming("")), Err(Err::Incomplete(Needed::new(5))));