Module nom8::character

Character specific parsers and combinators

Functions recognizing specific characters

Modules

complete

Character specific parsers and combinators, complete input version.

streaming

Character specific parsers and combinators, streaming version

Functions

alpha0

Recognizes zero or more lowercase and uppercase ASCII alphabetic characters: a-z, A-Z

alpha1

Recognizes one or more lowercase and uppercase ASCII alphabetic characters: a-z, A-Z

alphanumeric0

Recognizes zero or more ASCII numerical and alphabetic characters: 0-9, a-z, A-Z

alphanumeric1

Recognizes one or more ASCII numerical and alphabetic characters: 0-9, a-z, A-Z

crlf

Recognizes the string “\r\n”.

digit0

Recognizes zero or more ASCII numerical characters: 0-9

digit1

Recognizes one or more ASCII numerical characters: 0-9

f32

Recognizes floating point number in text format and returns a f32.

f64

Recognizes floating point number in text format and returns a f64.

hex_digit0

Recognizes zero or more ASCII hexadecimal numerical characters: 0-9, A-F, a-f

hex_digit1

Recognizes one or more ASCII hexadecimal numerical characters: 0-9, A-F, a-f

i8

will parse a number in text form to a number

i16

will parse a number in text form to a number

i32

will parse a number in text form to a number

i64

will parse a number in text form to a number

i128

will parse a number in text form to a number

line_ending

Recognizes an end of line (both ‘\n’ and ‘\r\n’).

multispace0

Recognizes zero or more spaces, tabs, carriage returns and line feeds.

multispace1

Recognizes one or more spaces, tabs, carriage returns and line feeds.

newline

Matches a newline character ‘\n’.

not_line_ending

Recognizes a string of any char except ‘\r\n’ or ‘\n’.

oct_digit0

Recognizes zero or more octal characters: 0-7

oct_digit1

Recognizes one or more octal characters: 0-7

recognize_float

Recognizes floating point number in a byte string and returns the corresponding slice.

recognize_float_parts

Recognizes a floating point number in text format

space0

Recognizes zero or more spaces and tabs.

space1

Recognizes one or more spaces and tabs.

tab

Matches a tab character ‘\t’.

u8

will parse a number in text form to a number

u16

will parse a number in text form to a number

u32

will parse a number in text form to a number

u64

will parse a number in text form to a number

u128

will parse a number in text form to a number