Module nom8::character::complete

Character specific parsers and combinators, complete input version.

Functions recognizing specific characters.

Functions

alpha0Deprecated

Recognizes zero or more lowercase and uppercase ASCII alphabetic characters: a-z, A-Z

alpha1Deprecated

Recognizes one or more lowercase and uppercase ASCII alphabetic characters: a-z, A-Z

alphanumeric0Deprecated

Recognizes zero or more ASCII numerical and alphabetic characters: 0-9, a-z, A-Z

alphanumeric1Deprecated

Recognizes one or more ASCII numerical and alphabetic characters: 0-9, a-z, A-Z

anycharDeprecated

Matches one byte as a character. Note that the input type will accept a str, but not a &[u8], unlike many other nom parsers.

charDeprecated

Recognizes one character.

crlfDeprecated

Recognizes the string “\r\n”.

digit0Deprecated

Recognizes zero or more ASCII numerical characters: 0-9

digit1Deprecated

Recognizes one or more ASCII numerical characters: 0-9

hex_digit0Deprecated

Recognizes zero or more ASCII hexadecimal numerical characters: 0-9, A-F, a-f

hex_digit1Deprecated

Recognizes one or more ASCII hexadecimal numerical characters: 0-9, A-F, a-f

i8

will parse a number in text form to a number

i16

will parse a number in text form to a number

i32

will parse a number in text form to a number

i64

will parse a number in text form to a number

i128

will parse a number in text form to a number

line_endingDeprecated

Recognizes an end of line (both ‘\n’ and ‘\r\n’).

multispace0Deprecated

Recognizes zero or more spaces, tabs, carriage returns and line feeds.

multispace1Deprecated

Recognizes one or more spaces, tabs, carriage returns and line feeds.

newlineDeprecated

Matches a newline character ‘\n’.

none_ofDeprecated

Recognizes a character that is not in the provided characters.

not_line_endingDeprecated

Recognizes a string of any char except ‘\r\n’ or ‘\n’.

oct_digit0Deprecated

Recognizes zero or more octal characters: 0-7

oct_digit1Deprecated

Recognizes one or more octal characters: 0-7

one_ofDeprecated

Recognizes one of the provided characters.

satisfyDeprecated

Recognizes one character and checks that it satisfies a predicate

space0Deprecated

Recognizes zero or more spaces and tabs.

space1Deprecated

Recognizes one or more spaces and tabs.

tabDeprecated

Matches a tab character ‘\t’.

u8

will parse a number in text form to a number

u16

will parse a number in text form to a number

u32

will parse a number in text form to a number

u64

will parse a number in text form to a number

u128

will parse a number in text form to a number