nom8::combinator
Function eof

pub fn eof<I, E: ParseError<I>>(
    input: I,
) -> IResult<I, <I as IntoOutput>::Output, E>
where
    I: InputLength + Clone + IntoOutput,

returns its input if it is at the end of input data

When we’re at the end of the data, this combinator will succeed

let parser = eof;
assert_eq!(parser("abc"), Err(Err::Error(("abc", ErrorKind::Eof))));
assert_eq!(parser(""), Ok(("", "")));