nom8::error
Function context

pub fn context<I: Clone, E: ContextError<I, &'static str>, F, O>(
    context: &'static str,
    f: F,
) -> impl FnMut(I) -> IResult<I, O, E>
where
    F: Parser<I, O, E>,

👎Deprecated since 8.0.0: Replaced with `Parser::context

Create a new error from an input position, a static string and an existing error. This is used mainly in the context combinator, to add user friendly information to errors when backtracking through a parse tree

WARNING: Deprecated, replaced with Parser::context