nom8::input
Trait HexDisplay

pub trait HexDisplay {
    // Required methods
    fn to_hex(&self, chunk_size: usize) -> String;
    fn to_hex_from(&self, chunk_size: usize, from: usize) -> String;
}

Helper trait to show a byte slice as a hex dump

Required Methods

fn to_hex(&self, chunk_size: usize) -> String

Converts the value of self to a hex dump, returning the owned String.

fn to_hex_from(&self, chunk_size: usize, from: usize) -> String

Converts the value of self to a hex dump beginning at from address, returning the owned String.