Trait nom8::input::InputIsStreaming

pub trait InputIsStreaming<const YES: bool>: Sized {
    type Complete: InputIsStreaming<false>;
    type Streaming: InputIsStreaming<true>;

    // Required methods
    fn into_complete(self) -> Self::Complete;
    fn into_streaming(self) -> Self::Streaming;
}

Marks the input as being the complete buffer or a partial buffer for streaming input

See Streaming for marking a presumed complete buffer type as a streaming buffer.

Required Associated Types

type Complete: InputIsStreaming<false>

Complete counterpart

Set to Self if this is a complete buffer.

Set to std::convert::Infallible if there isn’t an associated complete buffer type

type Streaming: InputIsStreaming<true>

Streaming counterpart

Set to Self if this is a streaming buffer.

Set to std::convert::Infallible if there isn’t an associated streaming buffer type

Required Methods

fn into_complete(self) -> Self::Complete

Convert to complete counterpart

fn into_streaming(self) -> Self::Streaming

Convert to streaming counterpart

Object Safety

This trait is not object safe.