Trait nom8::input::Location

pub trait Location {
    fn location(&self) -> usize;
}

Number of indices input has advanced since start of parsing

Required Methods

fn location(&self) -> usize

Number of indices input has advanced since start of parsing