nom8::input
Trait Offset

pub trait Offset {
    // Required method
    fn offset(&self, second: &Self) -> usize;
}

Useful functions to calculate the offset between slices and show a hexdump of a slice

Required Methods

fn offset(&self, second: &Self) -> usize

Offset between the first byte of self and the first byte of the argument

Object Safety

This trait is not object safe.