nom8::input
Trait UnspecializedInput

pub trait UnspecializedInput { }

Dummy trait used for default implementations (currently only used for InputTakeAtPosition and Compare).

When implementing a custom input type, it is possible to use directly the default implementation: If the input type implements InputLength, InputIter, InputTake and Clone, you can implement UnspecializedInput and get a default version of InputTakeAtPosition and Compare.

For performance reasons, you might want to write a custom implementation of InputTakeAtPosition (like the one for &[u8]).