Type Definition nom8::prelude::IResult

pub type IResult<I, O, E = Error<I>> = Result<(I, O), Err<E>>;

Holds the result of parsing functions

It depends on the input type I, the output type O, and the error type E (by default (I, nom8::ErrorKind))

The Ok side is a pair containing the remainder of the input (the part of the data that was not parsed) and the produced value. The Err side contains an instance of nom8::Err.

Outside of the parsing code, you can use the FinishIResult::finish method to convert it to a more common result type