nom8::bytes
Module streaming

Parsers recognizing bytes streams, streaming version

Functions

escapedDeprecated

Matches a byte string with escaped characters.

escaped_transformDeprecated

Matches a byte string with escaped characters.

is_aDeprecated

Returns the longest slice of the matches the pattern.

is_notDeprecated

Parse till certain characters are met.

tagDeprecated

Recognizes a pattern.

tag_no_caseDeprecated

Recognizes a case insensitive pattern.

takeDeprecated

Returns an input slice containing the first N input elements (Input[..N]).

take_tillDeprecated

Returns the longest input slice (if any) till a predicate is met.

take_till1Deprecated

Returns the longest (at least 1) input slice till a predicate is met.

take_untilDeprecated

Returns the input slice up to the first occurrence of the pattern.

take_until1Deprecated

Returns the non empty input slice up to the first occurrence of the pattern.

take_whileDeprecated

Returns the longest input slice (if any) that matches the predicate.

take_while1Deprecated

Returns the longest (at least 1) input slice that matches the predicate.

take_while_m_nDeprecated

Returns the longest (m <= len <= n) input slice that matches the predicate.