nom8::input
Trait AsChar

pub trait AsChar {
    // Required methods
    fn as_char(self) -> char;
    fn is_alpha(self) -> bool;
    fn is_alphanum(self) -> bool;
    fn is_dec_digit(self) -> bool;
    fn is_hex_digit(self) -> bool;
    fn is_oct_digit(self) -> bool;
    fn len(self) -> usize;
    fn is_space(self) -> bool;
    fn is_newline(self) -> bool;
}

Transforms common types to a char for basic token parsing

Required Methods

fn as_char(self) -> char

Makes a char from self

use nom8::input::AsChar as _;

assert_eq!('a'.as_char(), 'a');
assert_eq!(u8::MAX.as_char(), std::char::from_u32(u8::MAX as u32).unwrap());

fn is_alpha(self) -> bool

Tests that self is an alphabetic character

Warning: for &str it recognizes alphabetic characters outside of the 52 ASCII letters

fn is_alphanum(self) -> bool

Tests that self is an alphabetic character or a decimal digit

fn is_dec_digit(self) -> bool

Tests that self is a decimal digit

fn is_hex_digit(self) -> bool

Tests that self is an hex digit

fn is_oct_digit(self) -> bool

Tests that self is an octal digit

fn len(self) -> usize

Gets the len in bytes for self

fn is_space(self) -> bool

Tests that self is ASCII space or tab

fn is_newline(self) -> bool

Tests if byte is ASCII newline: \n